requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
Click==7.0
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
gunicorn==20.0.4
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
json5==0.9.2
jsonschema==3.2.0
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.4
Pygments==2.5.2
pyrsistent==0.15.7
python-dateutil==2.8.1
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.7.1
Send2Trash==1.5.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
APScheduler==3.0.0
xmltodict==0.12.0
python-jwt==3.2.6
gcloud==0.18.3
sseclient==0.0.24
requests-toolbelt==0.9.1
firebase-admin==4.0.0
